quiz/G1/Uzbakanov Omirzak/todoBack/requirements.txt
Django==2.2
django-cors-headers==2.5.2
django-filter==2.1.0
djangorestframework==3.9.2
pytz==2019.1
sqlparse==0.3.0
